feat(core): make test resource funcs  usable without test.TestSetup
